+++ doc/release-notes.md
@@ -3,3 +3,4 @@ Bitcoin ABC version 0.19.12 is now available from:
   <https://download.bitcoinabc.org/0.19.12/>
 
 This release includes the following features and fixes:
+  - Add the `getblockstats` RPC to get statistics on a block or a block range.